adding work-around for "unverified developer"
At the moment when a MacOS user (on at least macOS version 11) tries to open the app, they are unable to successfully open it.

This edit adds a "first launch" section to the readme for users who do not know how to fix this.

The way to fix this would be to register the app with Apple. When this happens this error won't occur.